functions/requirements.txt
azure-functions>=1.21.0
azure-identity>=1.19.0
msgraph-sdk>=1.12.0
python-dotenv>=1.0.0
